Code refactor for hyperphenotypes. I have decided that in the future i will implement more types of hyper substrates, so i have called this one LayeredHyperPhenotype
